add distance tracking, doesn't show on the app though

+++ app/src/main/java/com/demioshin/runway/data/RunData.kt
@@ -15,5 +15,5 @@ import kotlin.concurrent.thread
 class RunData {
     var time = MutableLiveData<Time>(Time(0, 0, 0))
     var steps = MutableLiveData<Int>(0)
-    var distance = MutableLiveData<Int>(0)
+    var distance = MutableLiveData<Float>(0F)
 }
